// frontend/src/app/models/anomaly-record.model.ts
export interface AnomalyRecord {
  id?: number;
  creType?: string;
  codApp?: string;
  batchid?: string;
  idCre?: string;
  codeErreur?: string;
  texteErreur?: string;
  nbrRecyclage?: string;
  darDate?: string;
  darTime?: string;
  codLot?: string;
  codAgence?: string;
  datOperat?: string;
  deviseOp?: string;
  mntOperat?: string;
  codeRecyclage?: string;
  codePhase?: string;
  codeDomaine?: string;
  texteCompErreur?: string;
  emetteur?: string;
  numCreErr?: string;
  codeCre?: string;
  versionCre?: string;
  codeInstance?: string;
  numAno?: string;
  typAno?: string;
  nivGene?: string;
  origineAno?: string;
  indEnregCre?: string;
  codeEnreg?: string;
  typeRegle?: string;
  codeRegleEnreg?: string;
  debVersionRegleEnreg?: string;
  finVersionRegleEnreg?: string;
  codeRegleMe?: string;
  debVersionRegleMe?: string;
  finVersionRegleMe?: string;
  code?: string;
  idfVacation?: string;
  idfEtape?: string;
  dateVacation?: string;
  heureVacation?: string;
  lot?: string;
  nivDetect?: string;
  codePrioSchema?: string;
  codeSchema?: string;
  numSeqGarn?: string;
  adrGarn?: string;
  codeFormatMe?: string;
  mnemoModule?: string;
  codeEtatAutom?: string;
  lgEnreg?: string;
  enreg?: string;
  fileName?: string;
}